+++ src/nl/underkoen/adventofcode/Day7.java
@@ -2,8 +2,8 @@ package nl.underkoen.adventofcode;
 
 import java.util.*;
 import java.util.concurrent.*;
-import java.util.function.IntConsumer;
-import java.util.function.IntSupplier;
+import java.util.function.LongConsumer;
+import java.util.function.LongSupplier;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
@@ -14,35 +14,35 @@ import static nl.underkoen.adventofcode.opcode.OpcodeRunner.process;
  * Created by Under_Koen on 06/12/2019.
  */
 public class Day7 extends AdventOfCode {
-    public static int processPhase(int[] program, Collection<Integer> phase) {
-        int out = 0;
-        for (Integer i : phase) out = process(program, new int[]{out, i});
+    public static long processPhase(long[] program, Collection<Integer> phase) {
+        long out = 0;
+        for (Integer i : phase) out = process(program, new long[]{out, i});
         return out;
     }
 
-    public static int processFeedbackPhase(int[] program, Collection<Integer> phase) {
-        List<BlockingQueue<Integer>> inputs = new ArrayList<>(5);
+    public static long processFeedbackPhase(long[] program, Collection<Integer> phase) {
+        List<BlockingQueue<Long>> inputs = new ArrayList<>(5);
 
         for (Integer i : phase) {
-            BlockingQueue<Integer> queue = new LinkedBlockingDeque<>();
-            queue.add(i);
+            BlockingQueue<Long> queue = new LinkedBlockingDeque<>();
+            queue.add((long) i);
             inputs.add(queue);
         }
-        inputs.get(0).add(0);
+        inputs.get(0).add(0L);
 
         ExecutorService pool = Executors.newFixedThreadPool(5);
         CountDownLatch latch = new CountDownLatch(5);
 
         for (int i = 0; i < 5; i++) {
             final int j = i;
-            IntSupplier inputF = () -> {
+            LongSupplier inputF = () -> {
                 try {
                     return inputs.get(j).take();
                 } catch (Exception e) {
                     return 0;
                 }
             };
-            IntConsumer output = (p) -> inputs.get((j + 1) % 5).add(p);
+            LongConsumer output = (p) -> inputs.get((j + 1) % 5).add(p);
             pool.submit(() -> {
                 process(program, inputF, output);
                 latch.countDown();
@@ -63,13 +63,13 @@ public class Day7 extends AdventOfCode {
     }
 
     @Override
-    public int[] getCorrectOutput() {
-        return new int[]{567045, 39016654};
+    public long[] getCorrectOutput() {
+        return new long[]{567045, 39016654};
     }
 
     @Override
     void run(List<String> input) {
-        int[] program = parse(input);
+        long[] program = parse(input);
         List<Set<Integer>> phases = IntStream.range(0, (int) Math.pow(5, 5))
                 .mapToObj(i -> {
                     Set<Integer> phase = new LinkedHashSet<>();
@@ -79,19 +79,19 @@ public class Day7 extends AdventOfCode {
                 .filter(phase -> phase.size() == 5)
                 .collect(Collectors.toList());
 
-        a = phases
+        a = (int) phases
                 .stream()
-                .mapToInt(phase -> processPhase(program, phase))
+                .mapToLong(phase -> processPhase(program, phase))
                 .max()
                 .orElse(0);
 
-        b = phases
+        b = (int) phases
                 .stream()
                 .map(phase -> phase.stream()
                         .map(i -> i + 5)
                         .collect(Collectors.toList())
                 )
-                .mapToInt(phase -> processFeedbackPhase(program, phase))
+                .mapToLong(phase -> processFeedbackPhase(program, phase))
                 .max()
                 .orElse(0);
     }
